Split feature-policy-frame-policy-allowed-for-some.https.sub.html
The wpt test was flaky because it has too much content: all 108
subtests live in a single test file, which causes the test sometimes
exceed the timeout limit(30s).

The solution to resolve this flakiness is to split this big test file
into smaller ones, so that each of them can have less content.

Bug: 1067391
Change-Id: Ib6ed1469895a76cff43c7e33dd24809242859044